+++ docs/changes.rst
@@ -12,6 +12,7 @@ all releases are available on `Anaconda.org <https://anaconda.org/pytask/pytask>
 - :gh:`55` implements miscellaneous fixes to improve error message, tests and coverage.
 - :gh:`59` adds a tutorial on using plugins and features plugins more prominently.
 - :gh:`60` adds the MIT license to the project and mentions pytest and its developers.
+- :gh:`63` adds versioneer to pytask.
 
 
 0.0.11 - 2020-12-27